banking-customer-service/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.0
pydantic==2.10.3
pydantic-settings==2.6.1
structlog==24.4.0
prometheus-fastapi-instrumentator==7.0.0
python-dotenv==1.0.1
pytest==8.3.4
pytest-asyncio==0.24.0
pytest-cov==6.0.0
factory-boy==3.3.1
httpx==0.28.1
email-validator==2.2.0
